+++ lockable/__init__.py
@@ -1,2 +1,3 @@
 """ Lockable module """
-from lockable.lockable import Lockable, ResourceNotFound, Allocation
+from lockable.lockable import Lockable, ResourceNotFound, Allocation, MODULE_LOGGER
+from lockable.provider import Provider, ProviderError
